requirements.txt
numpy
matplotlib
scipy
opensimplex
numpy-stl
